requirements.txt
aws-cdk-lib>=2.100.0
boto3>=1.28.0
pandas>=2.0.0
pyarrow>=12.0.0
pydantic>=2.0.0